Implemented Explorer in Home and Discover Screen & Handled SearchView Behavior

+++ app/src/main/java/com/example/newzz/repository/NewsRepository.kt
@@ -11,19 +11,24 @@ import com.example.newzz.db.ArticleDAO
 import com.example.newzz.model.Article
 import com.example.newzz.repository.paging.SearchNewsPagingSource
 import com.example.newzz.repository.paging.TopNewsPagingSource
+import com.example.newzz.util.ConnectivityObserver
+import com.example.newzz.util.NetworkChecker
 import kotlinx.coroutines.flow.Flow
 
 class NewsRepository(
     private val api: NewsAPI,
-    private val articleDAO: ArticleDAO
+    private val articleDAO: ArticleDAO,
+    private val networkChecker: NetworkChecker
 ) {
+
+    val isConnected = networkChecker.isNetworkAvailable()
 
     fun getTopNews(): Flow<PagingData<Article>> = Pager(
         config = PagingConfig(
             pageSize = 50,
             enablePlaceholders = false
         ),
-        pagingSourceFactory = { TopNewsPagingSource(api, articleDAO) }
+        pagingSourceFactory = { TopNewsPagingSource(api, articleDAO, networkChecker) }
     ).flow
 
     fun getSearches(query: String): Flow<PagingData<Article>> = Pager(
@@ -37,18 +42,22 @@ class NewsRepository(
     suspend fun getPopularNews(): List<Article> {
         Log.d("NewsRepository", "getPopularNews")
         return try {
-            val response = api.getTodayPopularNews()
-            Log.d("NewsRepository", "Response: ${response.message()}")
-            if (response.isSuccessful) {
-                val articles = response.body()?.articles?.filterNotNull() ?: emptyList()
-                Log.d("NewsRepository", "Articles: ${articles.size}")
-                val categorizedArticles = articles.map { it.copy(category = "today_popular") }
-                articleDAO.deleteArticlesByCategory("today_popular")
-                articleDAO.insertArticles(categorizedArticles)
-                categorizedArticles
+            if (isConnected) {
+                val response = api.getTodayPopularNews()
+                Log.d("NewsRepository", "Response: ${response.message()}")
+                if (response.isSuccessful) {
+                    val articles = response.body()?.articles?.filterNotNull() ?: emptyList()
+                    Log.d("NewsRepository", "Articles: ${articles.size}")
+                    val categorizedArticles = articles.map { it.copy(category = "today_popular") }
+                    articleDAO.deleteArticlesByCategory("today_popular")
+                    articleDAO.insertArticles(categorizedArticles)
+                    categorizedArticles
+                } else {
+                    Log.e("NewsRepository", "Error: ${response.message()}")
+                    emptyList()
+                }
             } else {
-                Log.e("NewsRepository", "Error: ${response.message()}")
-                emptyList()
+                articleDAO.getTodayPopularArticles()
             }
         } catch (e: Exception) {
             Log.e("NewsRepository", "Exception: ${e.message}")
@@ -58,19 +67,8 @@ class NewsRepository(
 
     fun getSavedArticles(): LiveData<List<Article>> = articleDAO.getSavedArticles()
 
-    fun getSearchedArticles(): LiveData<List<Article>> = articleDAO.getSearchedArticles()
-
-    fun getTopArticles(): LiveData<List<Article>> = articleDAO.getTopArticles()
-
-    fun getTodayPopularArticles(): LiveData<List<Article>> = articleDAO.getTodayPopularArticles()
-
     suspend fun saveStateChange(article: Article) {
         article.isSaved = !article.isSaved
         articleDAO.updateArticle(article)
-    }
-
-    suspend fun insertArticles(articles: List<Article>, category: String) {
-        val categorizedArticles = articles.map { it.copy(category = category) }
-        articleDAO.insertArticles(categorizedArticles)
     }
 }
